+++ Driver/Font/TrueType/FreeType/ttcache.h
@@ -110,8 +110,8 @@
   /* a cache kind.                                           */
   struct  TCache_Class_
   {
-    ULong         object_size;
-    Long          idle_limit;
+    UShort        object_size;
+    Short         idle_limit;
     PConstructor  init;
     PDestructor   done;
     PRefresher    reset;
@@ -146,7 +146,7 @@
     PCache_Class      clazz;      /* 'class' is a reserved word in C++ */
     TSingle_List      active;
     TSingle_List      idle;
-    Long              idle_count;
+    Short             idle_count;
   };
 
   typedef struct TCache_  TCache;
